pyproject.toml
[project]
name = "personal-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.6",
    "langgraph-checkpoint-postgres>=3.0.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
